+++ addon/components/flexberry-simpleolv.js
@@ -57,7 +57,7 @@ ErrorableControllerMixin, {
         this.set('contentWithKeys', this.contentForRender);
       }
 
-      this.set('showLoadingTbodyClass', false);
+      this.get('objectlistviewEventsService').setLoadingState('');
     }
   })),
 
@@ -457,15 +457,6 @@ ErrorableControllerMixin, {
   contentForRender: null,
 
   /**
-    Class loading for tbody.
-
-    @property showLoadingTbodyClass
-    @type Boolean
-    @defaul false
-  */
-  showLoadingTbodyClass: false,
-
-  /**
     Flag indicates whether content is defined.
 
     @property hasContent
@@ -723,7 +714,7 @@ ErrorableControllerMixin, {
         return;
       }
 
-      this.set('showLoadingTbodyClass', true);
+      this.get('objectlistviewEventsService').setLoadingState('loading');
 
       let action = e.ctrlKey ? 'addColumnToSorting' : 'sortByColumn';
       this.sendAction(action, column);
@@ -1071,15 +1062,6 @@ ErrorableControllerMixin, {
     this.get('colsConfigMenu').on('addNamedSetting', this, this.__addNamedSetting);
     this.get('colsConfigMenu').on('deleteNamedSetting', this, this._deleteNamedSetting);
 
-    let eventsBus = this.get('eventsBus');
-    if (eventsBus) {
-      eventsBus.on('showLoadingTbodyClass', (componentName, showLoadingTbodyClass) => {
-        if (componentName === this.get('componentName')) {
-          this.set('showLoadingTbodyClass', showLoadingTbodyClass);
-        }
-      });
-    }
-
     let projection = this.get('modelProjection');
 
     if (!projection) {
@@ -1177,10 +1159,6 @@ ErrorableControllerMixin, {
     this._super(...arguments);
 
     Ember.$(window).unbind(`resize.${this.get('componentName')}`);
-    let eventsBus = this.get('eventsBus');
-    if (eventsBus) {
-      eventsBus.off('showLoadingTbodyClass');
-    }
   },
 
   /**